fix(embeddings): address DGX feature review findings
