+++ src/engine/models/index.ts
@@ -10,11 +10,13 @@ export {
 export { createEmbeddingModel } from "./factory.js";
 export {
   getEmbeddingModelCatalogEntry,
+  listEmbeddingModels,
+} from "./catalog.js";
+export {
   getRemoteEmbeddingProviderCatalogEntry,
   isRemoteEmbeddingProvider,
-  listEmbeddingModels,
   listRemoteEmbeddingProviders,
-} from "./catalog.js";
+} from "../remote-embedding-providers.js";
 export {
   resolveEmbeddingReference,
   type ResolveEmbeddingReferenceOptions,
